requirements.txt
httpx==0.*
marshmallow-dataclass>=8.4.1
pyside2